refactor: cleanup folder
