+++ server/src/jobs/doc/generateTypes.js
@@ -1,0 +1,66 @@
+const fs = require("fs");
+const path = require("path");
+const { compile } = require("json-schema-to-typescript");
+const schemas = require("../../common/model/schema");
+const { getJsonFromMongooseSchema } = require("./mongooseUtils");
+
+const disableAdditionalProperties = (node) => {
+  if (node.properties) {
+    node.additionalProperties = false;
+    for (const key of Object.keys(node.properties)) {
+      disableAdditionalProperties(node.properties[key]);
+    }
+  }
+  if (node.items) {
+    disableAdditionalProperties(node.items);
+  }
+};
+
+const applyDateType = (node) => {
+  if (node.format === "date-time") {
+    node.tsType = "Date";
+  }
+
+  if (node.properties) {
+    for (const key of Object.keys(node.properties)) {
+      const field = node.properties[key];
+      applyDateType(field);
+    }
+  }
+
+  if (node.items) {
+    applyDateType(node.items);
+  }
+};
+
+const prepareJsonSchema = (jsonSchema) => {
+  const schema = { ...jsonSchema };
+
+  disableAdditionalProperties(schema);
+
+  applyDateType(schema);
+
+  return schema;
+};
+
+const generateTypes = () => {
+  Object.keys(schemas).forEach((schemaName) => {
+    const schema = schemas[schemaName];
+    const baseFilename = schemaName.replace("Schema", "");
+    const eJsonSchema = getJsonFromMongooseSchema(schema);
+    const jsonSchema = prepareJsonSchema(eJsonSchema);
+
+    // if (["formationSchema"].includes(schemaName)) {
+    //   let sout = {};
+    //   sout[baseFilename] = jsonSchema;
+    //   const edata = JSON.stringify(sout, null, 2);
+    //   const content = edata;
+    //   fs.writeFileSync(path.resolve(__dirname, `../../common/model/schema/${baseFilename}.jsonschema`), content);
+    // }
+    compile(jsonSchema, schemaName).then((ts) =>
+      fs.writeFileSync(path.resolve(__dirname, `../../common/model/schema/${baseFilename}.d.ts`), ts)
+    );
+  });
+};
+
+module.exports = { generateTypes };
